fix: moving register payment to the correct folder
