+++ src/http/use-cases/factories/make-register-payments-use-case.ts
@@ -1,8 +1,8 @@
-import { RegisterPaymentUseCase } from '@/http/register-payment-use-case';
+import { RegisterPaymentUseCase } from '@/http/use-cases/register-payment-use-case';
 import { MembersPrismaRepository } from '@/respositories/prisma/members-prisma-repository';
 import { PaymentsPrismaRepository } from '@/respositories/prisma/payments-prisma-repository';
 
-export function makeRegisterPaymentsUseCase(){
+export function makeRegisterPaymentsUseCase() {
     const membersRepository = new MembersPrismaRepository();
     const paymentsRepository = new PaymentsPrismaRepository();
     const registerPaymentUseCase = new RegisterPaymentUseCase(
